Move some codes to config.c
